iOS 11.3.1

+++ PrivateFrameworks/AdCore.framework/ADServer.h
@@ -3,12 +3,12 @@
  */
 
 @interface ADServer : ADSingleton <NSURLSessionDelegate> {
-    double  _configurationExpirationTime;
+    int  _configurationExpirationTime;
     NSDictionary * _configurations;
     NSURLSession * _session;
 }
 
-@property double configurationExpirationTime;
+@property int configurationExpirationTime;
 @property (retain) NSDictionary *configurations;
 @property (readonly, copy) NSString *debugDescription;
 @property (readonly, copy) NSString *description;
@@ -23,7 +23,7 @@
 - (void).cxx_destruct;
 - (void)URLSession:(id)arg1 task:(id)arg2 didCompleteWithError:(id)arg3;
 - (void)buildConfigurationDictionary:(id)arg1;
-- (double)configurationExpirationTime;
+- (int)configurationExpirationTime;
 - (id)configurations;
 - (void)handleRequest:(id)arg1 serverURL:(id)arg2 responseHandler:(id /* block */)arg3;
 - (id)init;
@@ -32,7 +32,7 @@
 - (id)serializeMessage:(id)arg1;
 - (id)serverURLForMessageClass:(Class)arg1 serverURL:(id)arg2;
 - (id)session;
-- (void)setConfigurationExpirationTime:(double)arg1;
+- (void)setConfigurationExpirationTime:(int)arg1;
 - (void)setConfigurations:(id)arg1;
 - (void)setSession:(id)arg1;
 - (bool)shouldRestoreConfiguration;
