+++ PrivateFrameworks/iWorkImport.framework/TNPageController.h
@@ -83,8 +83,8 @@
 - (id)hintCacheDictionary;
 - (void)i_invalidateHintCache;
 - (void)i_invalidatePageLayoutCache;
-- (id)i_layerForHeaderType:(int)arg1 fragment:(int)arg2 atPageCoordinate:(struct TSUCellCoord { unsigned short x1; unsigned char x2; unsigned char x3; })arg3;
-- (void)i_setLayer:(id)arg1 forHeaderType:(int)arg2 fragment:(int)arg3 atPageCoordinate:(struct TSUCellCoord { unsigned short x1; unsigned char x2; unsigned char x3; })arg4;
+- (id)i_layerForHeaderType:(long long)arg1 fragment:(long long)arg2 atPageCoordinate:(struct TSUCellCoord { unsigned short x1; unsigned char x2; unsigned char x3; })arg3;
+- (void)i_setLayer:(id)arg1 forHeaderType:(long long)arg2 fragment:(long long)arg3 atPageCoordinate:(struct TSUCellCoord { unsigned short x1; unsigned char x2; unsigned char x3; })arg4;
 - (bool)inDynamicContentScaleChange;
 - (id)initWithDocumentRoot:(id)arg1;
 - (void)invalidateCachedAutoFitContentScaleForSheet:(id)arg1;
