+++ protocols/TSWPHeaderFooterProvider.h
@@ -7,12 +7,12 @@
 
 - (double)bodyWidth;
 - (TSKDocumentRoot *)documentRoot;
-- (TSWPStorage *)headerFooter:(int)arg1 fragmentAtIndex:(int)arg2;
+- (TSWPStorage *)headerFooter:(long long)arg1 fragmentAtIndex:(long long)arg2;
 - (NSEnumerator *)headerFooterFragmentEnumerator;
-- (int)headerFooterTypeForModel:(id <TSKModel>)arg1;
-- (int)headerFragmentIndexForModel:(id <TSKModel>)arg1;
-- (bool)isHeaderFooterEmpty:(int)arg1;
-- (bool)isHeaderFooterEmpty:(int)arg1 fragmentAtIndex:(int)arg2;
+- (long long)headerFooterTypeForModel:(id <TSKModel>)arg1;
+- (long long)headerFragmentIndexForModel:(id <TSKModel>)arg1;
+- (bool)isHeaderFooterEmpty:(long long)arg1;
+- (bool)isHeaderFooterEmpty:(long long)arg1 fragmentAtIndex:(long long)arg2;
 - (bool)usesSingleHeaderFooter;
 
 @end
